requirements.txt
fastapi==0.110.1
uvicorn==0.30.6
httpx
dotenv-python
langgraph
langchain_fireworks
colorama
